requirements.txt
blinker==1.4
flask-login==0.3.2
flask-oauthlib==0.9.3
flask-restful==0.3.5
flask-wtf==0.13.1
flask==0.11.1
unidecode==0.4.19
webargs==1.3.4
GoogleAppEngineCloudStorageClient==1.9.22.1
